requirements.txt
# Core Requirements
streamlit
scikit-learn
pandas
numpy
joblib

# NLP Processing
nltk
regex

# Visualization (using compatible versions)
matplotlib
plotly
wordcloud 
Pillow

# System dependencies
python-dateutil
setuptools